+++ src/main/java/com/viettel/tuandz/service/dto/SysCatDTO.java
@@ -1,21 +1,17 @@
 package com.viettel.tuandz.service.dto;
 
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
 public class SysCatDTO {
 
-    private Integer id;
+    private Long id;
     private String code;
     private String name;
     private String description;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
